EN-865 Add dto for attribute validation objects
